Create navigation.api module
