+++ .idea/gradle.xml
@@ -25,6 +25,7 @@
             <option value="$PROJECT_DIR$/features/routines" />
             <option value="$PROJECT_DIR$/features/timer" />
             <option value="$PROJECT_DIR$/navigation" />
+            <option value="$PROJECT_DIR$/navigation/api" />
             <option value="$PROJECT_DIR$/ui_components" />
           </set>
         </option>
