+++ app/build.gradle
@@ -85,7 +85,7 @@ dependencies {
     implementation project(":features:timer")
     implementation project(":features:routines")
     implementation project(":ui_components")
-    implementation project(":navigation")
+    implementation project(":navigation:api")
 
     implementation Libs.AndroidX.coreKtx
     implementation Libs.AndroidX.appCompat
