+++ features/routines/src/main/java/com/enricog/routines/RoutinesNavigationGraph.kt
@@ -2,16 +2,16 @@ package com.enricog.routines
 
 import androidx.navigation.NavGraphBuilder
 import androidx.navigation.navigation
-import com.enricog.navigation.extensions.navViewModel
-import com.enricog.navigation.routes.RoutinesRoute
+import com.enricog.navigation.api.extensions.navViewModel
+import com.enricog.navigation.api.routes.RoutinesRoute
 import com.enricog.routines.detail.routine.RoutineScreen
 import com.enricog.routines.detail.segment.SegmentScreen
 import com.enricog.routines.detail.summary.RoutineSummaryScreen
 import com.enricog.routines.list.RoutinesScreen
-import com.enricog.navigation.routes.RoutineRoute.compose as composeRoutine
-import com.enricog.navigation.routes.RoutineSummaryRoute.compose as composeRoutineSummary
-import com.enricog.navigation.routes.RoutinesRoute.compose as composeRoutines
-import com.enricog.navigation.routes.SegmentRoute.compose as composeSegment
+import com.enricog.navigation.api.routes.RoutineRoute.compose as composeRoutine
+import com.enricog.navigation.api.routes.RoutineSummaryRoute.compose as composeRoutineSummary
+import com.enricog.navigation.api.routes.RoutinesRoute.compose as composeRoutines
+import com.enricog.navigation.api.routes.SegmentRoute.compose as composeSegment
 
 fun NavGraphBuilder.RoutinesNavigation() {
     navigation(
