+++ features/timer/build.gradle
@@ -49,7 +49,7 @@ dependencies {
     implementation project(":ui_components")
     implementation project(":entities")
     implementation project(":data:datasource")
-    implementation project(":navigation")
+    implementation project(":navigation:api")
 
     implementation Libs.AndroidX.appCompat
     implementation Libs.AndroidX.coreKtx
