+++ features/timer/src/main/java/com/enricog/timer/TimerNavigationGraph.kt
@@ -2,9 +2,9 @@ package com.enricog.timer
 
 import androidx.navigation.NavGraphBuilder
 import androidx.navigation.navigation
-import com.enricog.navigation.extensions.navViewModel
-import com.enricog.navigation.routes.TimerRoute
-import com.enricog.navigation.routes.TimerRoute.compose
+import com.enricog.navigation.api.extensions.navViewModel
+import com.enricog.navigation.api.routes.TimerRoute
+import com.enricog.navigation.api.routes.TimerRoute.compose
 
 fun NavGraphBuilder.TimerNavigation() {
     navigation(startDestination = TimerRoute.name, route = "timer") {
