+++ features/timer/src/main/java/com/enricog/timer/navigation/TimerNavigationActions.kt
@@ -1,9 +1,9 @@
 package com.enricog.timer.navigation
 
-import com.enricog.navigation.Navigator
-import com.enricog.navigation.routes.RoutinesRoute
-import com.enricog.navigation.routes.RoutinesRouteInput
-import com.enricog.navigation.routes.TimerRoute
+import com.enricog.navigation.api.Navigator
+import com.enricog.navigation.api.routes.RoutinesRoute
+import com.enricog.navigation.api.routes.RoutinesRouteInput
+import com.enricog.navigation.api.routes.TimerRoute
 import javax.inject.Inject
 
 internal class TimerNavigationActions @Inject constructor(
