+++ features/timer/src/test/java/com/enricog/timer/navigation/TimerNavigationActionsTest.kt
@@ -1,10 +1,10 @@
 package com.enricog.timer.navigation
 
 import com.enricog.base_test.coroutine.CoroutineRule
-import com.enricog.navigation.Navigator
-import com.enricog.navigation.routes.RoutinesRoute
-import com.enricog.navigation.routes.RoutinesRouteInput
-import com.enricog.navigation.routes.TimerRoute
+import com.enricog.navigation.api.Navigator
+import com.enricog.navigation.api.routes.RoutinesRoute
+import com.enricog.navigation.api.routes.RoutinesRouteInput
+import com.enricog.navigation.api.routes.TimerRoute
 import io.mockk.coVerify
 import io.mockk.mockk
 import org.junit.Rule
