+++ navigation/api/src/main/AndroidManifest.xml
