+++ navigation/api/src/main/java/com/enricog/navigation/api/NavigationAction.kt
@@ -1,0 +1,8 @@
+package com.enricog.navigation.api
+
+import androidx.navigation.NavOptions
+
+sealed class NavigationAction {
+    data class GoTo(val route: String, val navOptions: NavOptions?) : NavigationAction()
+    object GoBack : NavigationAction()
+}
