+++ navigation/api/src/main/java/com/enricog/navigation/api/Navigator.kt
@@ -1,0 +1,9 @@
+package com.enricog.navigation.api
+
+import kotlinx.coroutines.flow.Flow
+
+interface Navigator {
+    val actions: Flow<NavigationAction>
+
+    suspend fun navigate(action: NavigationAction)
+}
