+++ navigation/api/src/main/java/com/enricog/navigation/api/routes/RouteInput.kt
@@ -1,0 +1,3 @@
+package com.enricog.navigation.api.routes
+
+sealed interface RouteInput
