--- navigation/build.gradle
@@ -1,53 +1,0 @@
-import com.enricog.tempo.buildsrc.Libs
-import com.enricog.tempo.buildsrc.Versions
-
-plugins {
-    id 'com.android.library'
-    id 'kotlin-android'
-}
-
-android {
-    namespace 'com.enricog.navigation'
-
-    compileSdkVersion Versions.compileSdk
-
-    defaultConfig {
-        minSdkVersion Versions.minSdk
-    }
-
-    compileOptions {
-        sourceCompatibility JavaVersion.VERSION_1_8
-        targetCompatibility JavaVersion.VERSION_1_8
-    }
-
-    buildFeatures {
-        compose true
-    }
-    composeOptions {
-        kotlinCompilerVersion Libs.Kotlin.version
-        kotlinCompilerExtensionVersion Libs.AndroidX.Compose.version
-    }
-    packagingOptions {
-        jniLibs {
-            excludes += ['META-INF/licenses/**']
-        }
-        resources {
-            excludes += ['META-INF/licenses/**', 'META-INF/AL2.0', 'META-INF/LGPL2.1']
-        }
-    }
-}
-
-
-dependencies {
-    implementation project(":entities")
-    
-    implementation Libs.Kotlin.coroutines
-
-    implementation Libs.AndroidX.Compose.navigation
-    implementation Libs.AndroidX.Compose.runtime
-    implementation Libs.Hilt.navigation
-
-    testImplementation Libs.Kotlin.Test.coroutines
-    testImplementation Libs.Kotlin.Test.kotlin
-    testImplementation Libs.Test.junit
-}
