--- navigation/src/main/java/com/enricog/navigation/NavigationAction.kt
@@ -1,8 +1,0 @@
-package com.enricog.navigation
-
-import androidx.navigation.NavOptions
-
-sealed class NavigationAction {
-    data class GoTo(val route: String, val navOptions: NavOptions?) : NavigationAction()
-    object GoBack : NavigationAction()
-}
