--- navigation/src/main/java/com/enricog/navigation/Navigator.kt
@@ -1,9 +1,0 @@
-package com.enricog.navigation
-
-import kotlinx.coroutines.flow.Flow
-
-interface Navigator {
-    val actions: Flow<NavigationAction>
-
-    suspend fun navigate(action: NavigationAction)
-}
