--- navigation/src/main/java/com/enricog/navigation/routes/RouteInput.kt
@@ -1,3 +1,0 @@
-package com.enricog.navigation.routes
-
-sealed interface RouteInput
