+++ settings.gradle
@@ -9,4 +9,4 @@ include ':core'
 include ':entities'
 include ':data:datasource'
 include ':data:localdatasource'
-include ':navigation'
+include ':navigation:api'
